ARROW-185: Make padding and alignment for all buffers be 64 bytes
+ some small cleanup/removal of unnecessary code.  I think there is likely a good opportunity to factor this code better generally, but this seems to work for now.

Author: Micah Kornfield <emkornfield@gmail.com>

Closes #74 from emkornfield/emk_fix_allocations_PR and squashes the following commits:

e3cca14 [Micah Kornfield] fix cast style
1d006d8 [Micah Kornfield] fix warning
c140e04 [Micah Kornfield] fix lint
7543267 [Micah Kornfield] cleanup
11b3fd7 [Micah Kornfield] replace cython string conversion with string builder
05653cb [Micah Kornfield] add back in memsets because they make valgrind happy
6ff3048 [Micah Kornfield] ARROW-185: Make padding and alignment for all buffers be 64 bytes

+++ cpp/src/arrow/types/list.cc
@@ -47,7 +47,7 @@ bool ListArray::Equals(const std::shared_ptr<Array>& arr) const {
 Status ListArray::Validate() const {
   if (length_ < 0) { return Status::Invalid("Length was negative"); }
   if (!offset_buf_) { return Status::Invalid("offset_buf_ was null"); }
-  if (offset_buf_->size() / sizeof(int32_t) < length_) {
+  if (offset_buf_->size() / static_cast<int>(sizeof(int32_t)) < length_) {
     std::stringstream ss;
     ss << "offset buffer size (bytes): " << offset_buf_->size()
        << " isn't large enough for length: " << length_;
